V00717 - 2021-07-12 - Initial Upload
Work on the project has begun!

The theme is Death and it needs to be based on speculative play. As it stands, I'm going to attempt to go in the direction of a morality simulator through the lens of a fictional hospital death panel. I'd like to make a game where there DEFINETLY is an ending, but is it every really a good one?

So far, I'm only uploading a paper made prototype of in progress work. The environment will be adapted from a MacOS 6 UI for both a sense of outdated hospital equipment and to save on time when creating the aesthetic of the piece so I can focus more on mechanics.

On the subject of the mechanics, I have the following set out so far:

- Players will use a program called ammit.dmg to decide who recieves life saving surgery
- Players will comb through various other pieces of software to check on things like insurance coverage, organ availability, emails, and medical charts to find info on the patients
- Players need to maintain a profit level for the hospital and failure to do so will result in a game over
- Players only have so many organs for transplants
- Players will be given the moral decision making process over who gets the organs

The physical mechanics of players checking across multiples programs seems simple enough, but I'm still trying to figure out how to force the player to feel the full morality of their choices. I don't want this turning into a full resource manager, but rather a reflection of the players values that all factors including morality determine how they play.

Finally, I'm hoping to port over the Rainebow engine I have been working on into this project to not only continue it's development, but because it already contains a lot of what I'm looking to use including an animation system. Once I have the paper layout, I will begin work on prepping for the port.